Big change! Using code based on the little-known xzoom program, so no need for XFCE The big problem with XFCE's zoom was that it follwed the mouse. So there was no way to have the terminal representation of the desktop map to a smaller segment of the real desktop without the terminal mouse position being able to exactly 'hover' over the real mouse position.
The xzoom program is a window that displays a zoom of a portion of the desktop. So
double the width of the desktop and place the xzoom window on the right, but have
it watch only the half of the desktop on the left. What's more xzoom is small and
it's C code is easily incuded in the Golang code so they act as one, even sharing
state such as mouse coords, viewport position, current zoom level, etc.

WIP. Still contains old XFCE zoom code.