Fixes for text that clobbers other text
During the text builder process its possible that a character tries to
fit onto a TTY cell that already has a character. There are multiple
reasons for this, the main one being hidden text. This commit changes
the method for deciding the priority of character to using the
`elementFromPoint(x, y)` method.

This commit also includes a fix to re-parse all text when there's a
mousedown event.