Text rendering: fixes and improvements
Introduces a new option to disable the pixel comparison text visibility
detecting code. There are some situations in which it helps and some in
which it hinders. So for now it is disable by default and can be
enabled through the config file or my pressing F6 in the TTY client.

Also includes a couple of fixes to the HTTP server's text rendering.