fix: Don't load startup URL in HTTP Server mode
I think this is the source of a long-standing issue in the HTTP Server
where it would just freeze after a certain period of time. Every
successful page load, wether it was explicitly requested as a raw-text
request or not, attempts to send a raw text payload on page load. So I
think the automatic default home page startup loading was confusing
things.

Fixes #207